Add Password Hash & Verify Method

+++ cek_login.php
@@ -16,8 +16,8 @@ $total_count=$check_login_row['total_count'];
 if($total_count==3){
     echo "<script>alert('Terlalu banyak percobaan, coba lagi dalam 30 detik!'); window.location = 'index.php'</script>";
 }else{
-    $query = mysqli_prepare($koneksi, "SELECT * FROM tbl_admin WHERE username=? AND password=?");
-    mysqli_stmt_bind_param($query,"ss",$username,$password);
+    $query = mysqli_prepare($koneksi, "SELECT * FROM tbl_admin WHERE username=?");
+    mysqli_stmt_bind_param($query,"s",$username);
     mysqli_stmt_execute($query);
     $result=mysqli_stmt_get_result($query);
     $cek = mysqli_num_rows($result);    
@@ -25,20 +25,34 @@ if($total_count==3){
 
     if ($cek > 0) {
 
-        $_SESSION['namaadmin']  = $r['nama_admin'];
-        $_SESSION['username']   = $r['username'];
-        $_SESSION['password']   = $r['password'];
-        $_SESSION['idadmin']    = $r['id_admin'];
+        if(password_verify($password, $r['password'])){
+            $_SESSION['namaadmin']  = $r['nama_admin'];
+            $_SESSION['username']   = $r['username'];
+            $_SESSION['password']   = $r['password'];
+            $_SESSION['idadmin']    = $r['id_admin'];
 
-        if (!empty($_POST["remember"])) {
-            setcookie("username", $_POST["username"], time() + (60 * 60 * 24 * 5));
-            setcookie("password", $_POST["password"], time() + (60 * 60 * 24 * 5));
-        } else {
-            setcookie("username", "");
-            setcookie("password", "");
+            if (!empty($_POST["remember"])) {
+                setcookie("username", $_POST["username"], time() + (60 * 60 * 24 * 5));
+                setcookie("password", $_POST["password"], time() + (60 * 60 * 24 * 5));
+            } else {
+                setcookie("username", "");
+                setcookie("password", "");
+            }
+            mysqli_query($koneksi,"DELETE from login_logs WHERE ip_address='$ip_address'");
+            header("location:dashboard.php?hal=home");
         }
-        mysqli_query($koneksi,"DELETE from login_logs WHERE ip_address='$ip_address'");
-        header("location:dashboard.php?hal=home");
+        else{
+            $total_count++;
+            $rem_attm=3-$total_count;
+            if($rem_attm==0){
+                echo "<script>alert('Terlalu banyak percobaan, coba lagi dalam 30 detik!'); window.location = 'index.php'</script>";
+            }else{
+                echo "<script>alert('Password atau username salah! Sisa $rem_attm percobaan lagi!');window.location = 'index.php'</script>";
+            }
+            $try_time=time();
+            mysqli_query($koneksi,"INSERT into login_logs values('','$ip_address','$try_time')");
+        }
+
     } else {
         $total_count++;
         $rem_attm=3-$total_count;
